Use a better name for the first connexion
Use won't be able to "register" to wallabag but it'll have to "login" to the iOS app.
Using "register" user can understand it'll be able to register to use wallabag server/iOS/firefox/etc.. like creating an account.
Using "login" user might better understand it needs an account first to be able to use the iOS app