Fix for issue #339 with iOS app not syncing when articles are deleted from the web UI.
The previous fix does not solve this issue, hence submitting this one, which was more thoroughly tested. It turns out that the bug is actually an edge case: articles are not synced only when the list of articles becomes empty on the server. This is because `purge()` has a guard testing for `entriesSynced.count == 0` which prevents the sync from happening in the iOS app. This commit removes the guard, enabling the correct syncing behavior.

https://github.com/wallabag/ios-app/issues/339